// utils/types/requests/coworkingFormErrorInterface.ts
interface ICoworkingsErrorInfo {
  name?: string;
  lastname?: string;
  phone?: string;
  email?: string;
  identification?: string;
  position?: string;
  companyName?: string;
  companyEmail?: string;
  companyPhone?: string;
  address?: string;
  website?: string;
  open?: string;
  close?: string;
  capacity?: string;
  message?: string;
}

export default ICoworkingsErrorInfo;